move labels to folder
