+++ libs/template/labs/src/i18n/index.ts
@@ -1,6 +1,6 @@
 import { I18nFeature } from '@spryker-oryx/i18n';
 
-export * from './de';
+export * from './translations/de';
 
 // to avoid console errors, we provide a list of supported languages
 const supported = ['de'];
@@ -8,7 +8,7 @@ const supported = ['de'];
 export const labsI18nFeature = new I18nFeature({
   load: (localeId) => {
     if (supported.includes(localeId)) {
-      return import(`./${localeId}.ts`);
+      return import(`./translations/${localeId}.ts`);
     } else {
       return Promise.resolve();
     }
