+++ libs/template/labs/src/i18n/translations/de.ts
